+++ Sources/Shared/Toolkit/Closeable.swift
@@ -10,6 +10,7 @@ import Foundation
 public protocol Closeable {
     /// Closes this object and releases any resources associated with it.
     /// If the object is already closed then invoking this method has no effect.
+    @available(*, deprecated, message: "Handle Resource deallocation with `deinit` instead.")
     func close()
 }
 
@@ -20,6 +21,7 @@ public extension Closeable {
 public extension Closeable {
     /// Executes the given block function on this resource and then closes it down correctly whether
     /// an error is thrown or not.
+    @available(*, deprecated, message: "The resource is automatically closed when deallocated")
     @inlinable func use<T>(_ block: (Self) throws -> T) rethrows -> T {
         // Can't use `defer` with async functions.
         do {